requirements.txt
minio
dbt-snowflake
pandas
requests
python-dotenv
snowflake-connector-python[pandas]
dbt-core
dbt-snowflake